+++ common/util.h
@@ -2,29 +2,28 @@
 #define UTIL_H
 
 #include <assert.h>
-#include <stdint.h>
 #include <stdlib.h>
 #include "COMPARE.h"
 
 // rand returns a random index for an array of the specified size.
 // This is used to get the random pivot for use in QuickSelect.
-uint16_t randIndex(uint16_t size);
+int randIndex(int size);
 
 // swap modifies the given array by swapping two elements in place.
-void swap(uint16_t arr[], uint16_t a, uint16_t b);
+void swap(int arr[], int a, int b);
 
 // max chooses the largest of the two specified numbers.
-uint16_t max(uint16_t a, uint16_t b);
+int max(int a, int b);
 
 // min chooses the smallest of the two specified numbers.
-uint16_t min(uint16_t a, uint16_t b);
+int min(int a, int b);
 
 // debug prints out the array specified
-void debug(uint16_t arr[], uint16_t start, uint16_t end);
+void debug(int arr[], int start, int end);
 
 // shift modifies the given array to move the specified element to the left
 // or right a number of times. This shifts the rest of the elements around.
-void shift(uint16_t arr[], uint16_t pos, int shiftAmount);
+void shift(int arr[], int pos, int shiftAmount);
 
 // For finding # of comparisons
 int getComps(void);
